requirements.txt
alembic==1.8.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
Flask==2.1.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
GeoAlchemy2==0.12.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-editor==1.0.4
pytz==2019.3
pyxdg==0.26
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
WTForms==3.0.1
